+++ TODO.MD
@@ -1,0 +1,6 @@
+# Confront source code with following DDD patterns:
+## modules (done)
+## Integrating contexts
+## domain services
+## application services
+## architecture
