+++ actracker-search-feeder-main/src/main/java/ovh/equino/actracker/searchfeed/domain/model/activity/ActivityStore.java
@@ -1,4 +1,4 @@
-package ovh.equino.actracker.searchfeed.domain.activity;
+package ovh.equino.actracker.searchfeed.domain.model.activity;
 
 import java.util.Optional;
