+++ actracker-search-feeder-main/src/main/java/ovh/equino/actracker/searchfeed/infrastructure/index/console/ConsoleActivityIndex.java
@@ -1,0 +1,12 @@
+package ovh.equino.actracker.searchfeed.infrastructure.index.console;
+
+import ovh.equino.actracker.searchfeed.domain.model.activity.Activity;
+import ovh.equino.actracker.searchfeed.domain.model.activity.ActivityIndex;
+
+class ConsoleActivityIndex implements ActivityIndex {
+
+    @Override
+    public void index(Activity activity) {
+        System.out.printf("Indexed activity: %s%n", activity.id());
+    }
+}
