+++ actracker-search-feeder-main/src/main/java/ovh/equino/actracker/searchfeed/main/springboot/config/ApplicationConfig.java
@@ -3,7 +3,7 @@ package ovh.equino.actracker.searchfeed.main.springboot.config;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
-import ovh.equino.actracker.searchfeed.application.activity.ActivityService;
+import ovh.equino.actracker.searchfeed.application.ActivityService;
 
 @Configuration
 @ComponentScan(value = "ovh.equino.actracker.searchfeed.application", includeFilters = {
