+++ actracker-search-feeder-main/src/main/java/ovh/equino/actracker/searchfeed/main/springboot/config/index/ConsoleActivityIndexConfig.java
@@ -3,7 +3,7 @@ package ovh.equino.actracker.searchfeed.main.springboot.config.index;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
-import ovh.equino.actracker.searchfeed.domain.activity.ActivityIndex;
+import ovh.equino.actracker.searchfeed.domain.model.activity.ActivityIndex;
 
 @Configuration
 @ComponentScan(value = "ovh.equino.actracker.searchfeed.infrastructure.index.console", includeFilters = {
